move: 패키지 구조 변경
com
 ㄴ alumininum
     ㄴ acoe
         ㄴ com
             ㄴ auth
             ㄴ common
             ㄴ config
             ㄴ exception
             ㄴ api
             ㄴ util
			 ㄴ   ...
		 ㄴ domain
         |   ㄴ map
         |   |   ㄴ controller
         |   |   ㄴ service
         |   |   ㄴ dto
         |   |   ㄴ repository
         |   ...
