+++ src/main/java/com/aluminium/acoe/com/config/CorsConfig.java
@@ -1,4 +1,4 @@
-package com.aluminium.acoe.config;
+package com.aluminium.acoe.com.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
