+++ src/main/java/com/aluminium/acoe/com/entity/Authority.java
@@ -1,4 +1,4 @@
-package com.aluminium.acoe.entity;
+package com.aluminium.acoe.com.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
