+++ src/main/java/com/aluminium/acoe/com/exception/DuplicateMemberException.java
@@ -1,4 +1,4 @@
-package com.aluminium.acoe.exception;
+package com.aluminium.acoe.com.exception;
 
 public class DuplicateMemberException extends RuntimeException {
     public DuplicateMemberException() {
