+++ src/main/java/com/aluminium/acoe/com/jwt/JwtAuthenticationEntryPoint.java
@@ -1,4 +1,4 @@
-package com.aluminium.acoe.jwt;
+package com.aluminium.acoe.com.jwt;
 
 import org.springframework.security.core.AuthenticationException;
 import org.springframework.security.web.AuthenticationEntryPoint;
